ai_service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
